fix velocity command

+++ include/diffdrive_arduino/arduino_comms.h
@@ -20,7 +20,7 @@ public:
   void setup(const std::string &serial_device, int32_t baud_rate, int32_t timeout_ms);
   void sendEmptyMsg();
   void readEncoderValues(int &val_1, int &val_2);
-  void setMotorValues(int val_1, int val_2);
+  void setMotorValues(float val_1, float val_2);
   void setPidValues(float k_p, float k_d, float k_i, float k_o);
 
   bool connected() const { return serial_conn_.isOpen(); }
